+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/BranchDao.java
@@ -2,6 +2,5 @@ package com.opentravelsoft.providers;
 
 import com.opentravelsoft.entity.Branch;
 
-public interface BranchDao extends GenericDao<Branch, Long>
-{
+public interface BranchDao extends GenericDao<Branch, Integer> {
 }
