+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/CategoryDao.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import com.opentravelsoft.entity.Category;
 
-public interface CategoryDao extends GenericDao<Category, Long>
-{
-    List<Category> getCategory();
+public interface CategoryDao extends GenericDao<Category, Integer> {
+  List<Category> getCategory();
 }
