+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/ContactDao.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 import com.opentravelsoft.entity.Contact;
 
-public interface ContactDao extends GenericDao<Contact, Long> {
+public interface ContactDao extends GenericDao<Contact, Integer> {
   /**
    * 取得客户的联系人
    * 
    * @param accountId 客户ID
    * @return
    */
-  public List<Contact> searchContact(long accountId, boolean active);
+  public List<Contact> searchContact(int accountId, boolean active);
 
 }
