+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/EmployeeDao.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 import com.opentravelsoft.entity.Employee;
 
-public interface EmployeeDao extends GenericDao<Employee, Long> {
+public interface EmployeeDao extends GenericDao<Employee, Integer> {
   public List<Employee> getEmployees(boolean enabled);
 
   /**
@@ -14,7 +14,7 @@ public interface EmployeeDao extends GenericDao<Employee, Long> {
    * @param userId
    * @return
    */
-  public Employee getEmployee(long userId);
+  public Employee getEmployee(int userId);
 
   /**
    * 
@@ -37,7 +37,7 @@ public interface EmployeeDao extends GenericDao<Employee, Long> {
    * @param userId
    * @return
    */
-  public int deleteEmployee(long userId);
+  public int deleteEmployee(int userId);
 
   public int insertEmployee(Employee employee);
 
@@ -53,11 +53,11 @@ public interface EmployeeDao extends GenericDao<Employee, Long> {
 
   public int updateEmployee(Employee employee);
 
-  public Set<String> getAuthorities(long userId);
+  public Set<String> getAuthorities(int userId);
 
   public List<Employee> getSalesmans(boolean active);
 
-  public List<Employee> getSalesByTeam(long teamId);
+  public List<Employee> getSalesByTeam(int teamId);
 
-  public List<Employee> getUserByTeam(long teamId);
+  public List<Employee> getUserByTeam(int teamId);
 }
