+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/GroupDao.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import com.opentravelsoft.entity.Group;
 
-public interface GroupDao extends GenericDao<Group, Long>
-{
-    public List<Group> getAllGroups();
+public interface GroupDao extends GenericDao<Group, Integer> {
+  public List<Group> getAllGroups();
 }
