+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/TourCostDao.java
@@ -2,6 +2,6 @@ package com.opentravelsoft.providers;
 
 import com.opentravelsoft.entity.TourCost;
 
-public interface TourCostDao extends GenericDao<TourCost, Long> {
+public interface TourCostDao extends GenericDao<TourCost, Integer> {
 
 }
